Query Based Program to launch Applications from the OS
This is a solution to the challenge given by Mr Vimal Daga sir as part of the IIEC Rise Python Program.
This code is as per a Windows machine and only has generic applications that we commonly use. It can be made more versatile and flexible by including more applications but that would be dependent on the user.